Module Cobol_parser

Exported modules

module Options : sig ... end

Options to tune the parser engine

module Outputs : sig ... end

Output types for the engine

module Diagnostics : sig ... end

module Tokens : sig ... end

module Keywords : sig ... end

Exported functions

Basic (one-shot) parsing

type 'm simple_parsing = options:Cobol_parser__.Parser_options.parser_options -> Cobol_preproc.preprocessor -> (Cobol_ptree.compilation_group option, 'm) Cobol_parser__.Parser_outputs.output Cobol_common__Diagnostics_accumulator.MAKE(Cobol_parser__.Parser_diagnostics.ALL).TYPES.with_diags

Simple parsing functions traverse the inputs once to produce a result.

val parse_simple : Cobol_common.Behaviors.amnesic simple_parsing

Simple parsing function that does not return any artifact.

val parse_with_artifacts : Cobol_common.Behaviors.eidetic simple_parsing

Simple parsing function does return some artifacts.

Rewindable parsing

type 'm rewindable_parsing = options:Cobol_parser__.Parser_options.parser_options -> Cobol_preproc.preprocessor -> ((Cobol_ptree.compilation_group option, 'm) Cobol_parser__.Parser_outputs.output as 'x * 'x rewinder) Cobol_common__Diagnostics_accumulator.MAKE(Cobol_parser__.Parser_diagnostics.ALL).TYPES.with_diags

Rewindable parsing functions extend the behaviors of simple parsing functions, with the ability to rewind the parser (and pre-processor) before a given lexing position. To this end, their results include a rewinder that may then be given to rewind_and_parse.

and 'x rewinder

Rewinder for parsing functions that produce results of type 'x

and preprocessor_rewind = ?new_position:Stdlib.Lexing.position -> Cobol_preproc.preprocessor as 'r -> 'r

Functions for rewinding the pre-processor. Such a function should return a preprocessor in the exact same state as the one given in argument, with the only exception that the input text is now read from new_position. If new_position is not given, the text should be read from the very begining of the input.

val rewindable_parse_simple : Cobol_common.Behaviors.amnesic rewindable_parsing

Rewindable parsing function that does not return any artifact.

val rewindable_parse_with_artifacts : Cobol_common.Behaviors.eidetic rewindable_parsing

Rewindable parsing function that does return some artifacts.

type position =

| Lexing of Stdlib.Lexing.position
(*
raw lexing position
*)

| Indexed of {

line : int;
(*
line number (starting from 0)
*)

char : int;
(*
character number in line (starting from 0)
*)

}

Specification of positions to rewind to.

val rewind_and_parse : 'x rewinder -> preprocessor_rewind -> position:position -> ((Cobol_ptree.compilation_group option, 'm) Cobol_parser__.Parser_outputs.output as 'x * 'x rewinder) Cobol_common__Diagnostics_accumulator.MAKE(Cobol_parser__.Parser_diagnostics.ALL).TYPES.with_diags

rewind_and_parse rewinder preprocessor_rewind ~position uses rewinder to restart parsing before the given position. Note that the new_position argument that is given to preprocessor_rewind may not correspond to position, as the parser typically needs to rewind earlier than position.

Accessing artifacts

val artifacts : (_, Cobol_common.Behaviors.eidetic) Cobol_parser__.Parser_outputs.output -> Cobol_parser__.Parser_outputs.artifacts

Modules and functions exported for testing purposes

Signatures of modules below may change unexpectedly.

module INTERNAL : sig ... end