Module Cobol_parser.Outputs

Output types for the engine

Gathers some types used to define outputs for the parser engine.

type tokens_with_locs = Cobol_parser__.Grammar_tokens.token Cobol_common.Srcloc.TYPES.with_loc list

type artifacts = {

tokens : tokens_with_locs Stdlib.Lazy.t;

pplog : Cobol_preproc.Trace.log;

rev_comments : Cobol_preproc.Text.comments;

rev_ignored : Cobol_common.Srcloc.TYPES.lexloc list;

}

type ('result, 'memo) output =

| Only : 'result -> ('result, Cobol_common.Behaviors.amnesic) output

| WithArtifacts : 'result * artifacts -> ('result, Cobol_common.Behaviors.eidetic) output

The output of parsing functions depends on its memorization abilities:

an amnesic parse (when 'memo = Cobol_common.Behaviors.amnesic) only returns a parsing result (of type 'result);

an eidetic parse (when 'memo = Cobol_common.Behaviors.eidetic) additionally returns some parsing artefacts.

type 'm parsed_compilation_group = (Cobol_ptree.compilation_group option, 'm) output